fix(LottoMakeService):PurchaseAmount의 타입과 로또 생성에서 사용되는 count 변수 타입이 일치하지 않는 문제 수정

+++ src/test/java/lotto/model/LottoMakeServiceTest.java
@@ -30,7 +30,7 @@ public class LottoMakeServiceTest {
     @DisplayName("로또 여러장 생성 서비스 테스트")
     void testLottoMakeMultipleService() {
         LottoMakeService lottoMakeService = new LottoMakeService();
-        int count = 5;
+        Long count = 5L;
         LottoRepository lottoRepository = lottoMakeService.makeLottos(count);
 
         assertEquals(lottoRepository.getLottos().size(), count);
